requirements.txt
# For testing django support
Django

# Unit testing tools
nose
mock

# Haystack, Whoosh
django-haystack
Whoosh

# Docs
sphinx

# Dates
pytz
python-dateutil
